% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_tract_to_zcta_fun.R
\name{acs_tract_to_zcta_fun}
\alias{acs_tract_to_zcta_fun}
\title{Approximate ZCTA-level ACS data with census tracts}
\usage{
acs_tract_to_zcta_fun(state_num, acs_vars, acs_year)
}
\arguments{
\item{state_num}{A 2-digit state FIPS code. Both numeric and character
values are accepted. ACS data can only be downloaded for one state at a time.}

\item{acs_vars}{A vector of ACS 5-year survey variables.}

\item{acs_year}{A numeric year which indicates the final year in the 5-year
ACS survey data (e.g., \code{acs_year = 2015} results in 2011-2015 ACS data).
The value must be between 2009 and 2019.}
}
\value{
A data frame containing the best approximation of ZCTA-level ACS data
for the state and year provided. The data will contain one row per
combination of ZCTA & ACS variable. The data frame will contain the following
columns:
\itemize{
\item \code{state_fips} - (character) 2-digit state FIPS code, matching the provided
value of \code{state_num}.
\item \code{zcta} - (character) 5-digit ZCTA ID code
\item \code{variable} - (character) the ACS variable(s) requested.
\item \code{est_zcta} - (numeric) the ZCTA-level estimate of the ACS variable.
\item \code{p_zcta_covered} - (numeric) the proportion (valued 0-1) of the ZCTA
(weighted by population) that was able to be matched with census tracts and,
in essense, "covered" by the aggregated full and partial tract-level
estimates. For example, a value of 0.5 indicates that only 50\% of the ZCTA is
represented by this approximated value.
}
}
\description{
Approximate ZCTA (ZIP Code Tabulation Area) level ACS variables using
available census-tract level ACS data, for any year between 2009 and 2019.
This process utilizes both the tidycensus package and a tract-to-ZCTA
crosswalk.
}
\details{
ACS 5-year survey variables generally begin with "B". Note that the
tidycensus format for ACS variables includes an underscore between the
table and variable number, e.g., "B01001_001". For a full list of
ACS 5-year survey variable names, see tidycensus::load_variables().

This process only works for count data. Do not use this function for
non-count variables (medians, proportions, etc.).

Data source used for ZCTA-tract crosswalks:
Missouri Census Data Center: Geographic Correspondence Engine.
Web page: https://mcdc.missouri.edu/applications/geocorr.html

If you are new to tidycensus, and need to set up an API key for census data,
see this help page to get started:
https://walker-data.com/tidycensus/articles/basic-usage.html
}
\examples{
temp_pop_ny_2009 <- acs_tract_to_zcta_fun("36","B01001_001",2009)
temp_pop_ny_pa_2009 <- do.call(rbind, lapply(c("36","42"),function(i){
  acs_tract_to_zcta_fun(i,"B01001_001",2009)}))

}
