% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_zcta_demo_fun.R
\name{approx_zcta_demo_fun}
\alias{approx_zcta_demo_fun}
\title{Approximate ZCTA-level summary demographic data via tract-level ACS data}
\usage{
approx_zcta_demo_fun(
  state_num = NULL,
  state_abbr = NULL,
  acs_year,
  all_demo = F,
  age = F,
  race = F,
  race_hisp = F,
  poverty = F
)
}
\arguments{
\item{state_num}{A 2-digit state FIPS code. Both numeric and character
values are accepted. ACS data can only be downloaded for one state at a time.
If NULL, \code{state_abbr} will be used instead.}

\item{state_abbr}{A 2-letter state abbreviation. If NULL, \code{state_num} will
be used instead.}

\item{acs_year}{A numeric year which indicates the final year in the 5-year
ACS survey data (e.g., \code{acs_year = 2015} results in 2011-2015 ACS data).
The value must be between 2009 and 2019.}

\item{all_demo}{Boolean (T/F). If T, data for all demographic categories
(see below) will be returned.}

\item{age}{Boolean (T/F). If T, the returned data will include the
population distribution of 5-year age groups (from ACS table B01001).}

\item{race}{Boolean (T/F). If T, the returned data will include the
population distribution of race groups  (from ACS table B02001).}

\item{race_hisp}{Boolean (T/F). If T, the returned data will include the
population distribution of race/Hispanic ethnicity groups
(from ACS table B03002).}

\item{poverty}{Boolean (T/F). If T, the returned data will include the
proportion of the population below poverty level (from ACS table B17001).}
}
\value{
A data frame containing the best approximation of ZCTA-level ACS data
for the state and year provided. The data will contain one row per ZCTA.
A selection of summary demographic data will be included in the returned
data. By default, all data includes a population total. Other proportion
variables (valued 0-1) may be included according to parameters above.
}
\description{
Approximate a selection of ZCTA (ZIP Code Tabulation Area) level summary
demographic data sourced from tract-level ACS data, for any year between
2009 and 2019. This process utilizes both the tidycensus package and a
tract-to-ZCTA crosswalk.
}
\details{
Data source used for ZCTA-tract crosswalks:
Missouri Census Data Center: Geographic Correspondence Engine.
Web page: https://mcdc.missouri.edu/applications/geocorr.html

If you are new to tidycensus, and need to set up an API key for census data,
see this help page to get started:
https://walker-data.com/tidycensus/articles/basic-usage.html

See also documentation for HERcrosswalks::acs_tract_to_zcta_fun()
}
\examples{
ny_demo_2009 <-
  approx_zcta_demo_fun(state_abbr = "NY", acs_year = 2009, all_demo = TRUE)
pa_de_ages_2009 <-
  do.call(rbind, lapply(c("42", "10"),function(i){
    approx_zcta_demo_fun(state_num = i, acs_year = 2009, age = TRUE)}))
de_race_2010 <-
  approx_zcta_demo_fun(state_abbr = "DE", acs_year = 2010, race = TRUE)
ak_raceeth_2011 <-
  approx_zcta_demo_fun(state_num = 2, acs_year = 2011, race_hisp = TRUE)
de_pov_2019 <-
  approx_zcta_demo_fun(state_num = 10, acs_year = 2019, poverty = TRUE)
}
