% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_county_proxies_fun.R
\name{city_county_proxies_fun}
\alias{city_county_proxies_fun}
\title{Determine county proxies for the largest cities in the US}
\usage{
city_county_proxies_fun(n_cities = 50, year, rule, min_p_county, min_p_city)
}
\arguments{
\item{n_cities}{The number of large cities for which county proxies are
desired. Defaults to 50.}

\item{year}{The year of population data used to rank city sizes and
determine the largest \code{n_cities} cities.}

\item{rule}{The rule used for determining whether a county should be
considered an appropriate proxy for the city. Options are \code{"or"}, \code{"and"},
\code{"county"}, and \code{"city"}.
\itemize{
\item \code{"county"}: Only the \code{min_p_county} requirement must be satisfied.
\code{min_p_city} is not considered.
\item \code{"city"}: Only the \code{min_p_city} requirement is considered.
\code{min_p_county} is not considered.
\item \code{"and"}: Both \code{min_p_county} and \code{min_p_city} requirements must be
satisfied.
\item \code{"or"}: Either the \code{min_p_county} requirement or \code{min_p_city}
requirement must be satisfied.
}}

\item{min_p_county}{The minimum proportion of the county that must
overlap with the city (weighted by population) for the county to be considered
an appropriate proxy for the city. Proportions are valued 0.0001 (0.01\%)
to 1 (100\%).}

\item{min_p_city}{The minimum proportion of the city that must
overlap with the county (weighted by population) for the county to be
considered an appropriate proxy for the city. Proportions are valued
0.0001 (0.01\%) to 1 (100\%).}
}
\value{
A data frame containing a crosswalk between cities and
appropriate county proxies. The data frame will contain the following
columns:
\itemize{
\item \code{rank} - (numeric) rank of the city by population size. The largest city has
a rank of 1
\item \code{city_name} - (character) the name of the city or place
\item \code{state_abbr} (character) - the 2-letter state abbreviation for the city
\item \code{city_fips} - (character) the 7-digit FIPS code for the city, created by
concatenating the 2-digit state FIPS code and 5-digit place FIPS code
\item \code{county_name} - (character) the county name (with state abbreviation)
for the city's county proxy
\item \code{county_fips} - (character) the 5-digit FIPS code for the county, created by
concatenating the 2-digit state FIPS code and 3-digit county FIPS code
\item \code{p_city} - (numeric) the proportion of the city that overlaps with the
county (weighted by population), valued 0 - 1
\item \code{p_county} - (numeric) the proportion of the county that overlaps with the
city (weighted by population), valued 0 - 1
\item \code{city_pop_est} - (numeric) the population estimate for the city in
the \code{year} specified in the input
}
}
\description{
Create a data set containing county proxy information for the largest
cities in the US, according to custom city-county match criteria.
}
\details{
For this data, cities are defined as places, which can be incorporated
places (such as cities, towns, or villages), or census designated places
(CDPs).

"Counties" includes several types of county-equivalent entities, such as:
\itemize{
\item Parishes (Louisiana)
\item Independent cities (Maryland, Missouri, Nevada, Virginia)
\item District of Columbia (Washington, DC)
\item Boroughs, municipalities, and census areas (Alaska)
}

Cities in the US often do not cleanly follow the boundaries of US counties.
The following situations are possible for a city:
\itemize{
\item the city is contained wholly within one county, with a remaining
portion of the county not belonging to that city;
\item the city is equivalent to one or more counties, e.g., the 5 boroughs of
New York, NY which are all distinct counties;
\item the city overlaps with portions of multiple counties. In these cases,
a portion of a county that overlaps with a certain city may potentially be
very small, or vice versa.
}
\subsection{Limitations of the Data}{

During the process to create the requested data, cities will be sorted by
population size before applying the \code{min_p_city} and/or \code{min_p_county} rules.
Some cities may be eliminated from the data entirely if no
appropriate county matches are found which satisfy the user's conditions,
especially for large values of \code{min_p_city} and/or \code{min_p_county}.
If that happens, the number of cities in the final resulting crosswalk may
be less than the requested n_cities.

Over time, some counties and cities/towns change names, split into multiple
parts, or merge into other areas. In these cases, not only will the name of
a county or place change, but so will its geographic ID/FIPS code.

Due to limitations of the source data, some cities which ceased to
exist (in their original forms) prior to 2019 may not be included in the
data at all, or may be listed as different (post-2019 equivalent) names or
FIPS codes than what would have existed at an earlier time. User-generated
crosswalks for earlier years may include some cities which did not exist yet
in those earlier years.

Additionally, the county proxies returned in the requested crosswalk data
may be different according to which \code{year} the user inputs. Requests for
years 2010-2019 utilize a 2010 county-place relationship file, while requests
for years 2020 onward utilize a 2020 county-place relationship file.
}

\subsection{Data sources used:}{

American National Standards Institute (ANSI) and Federal Information
Processing Series (FIPS) Codes: 2010 ANSI Codes for Places.
Web page: https://www.census.gov/library/reference/code-lists/ansi.html#place

Relationship Files: State-level 2020 Place to 2010 Place Relationship Files.
Web page: https://www.census.gov/geographies/reference-files/time-series/geo/relationship-files.2020.html#place

Missouri Census Data Center: Geographic Correspondence Engine.
Web page: https://mcdc.missouri.edu/applications/geocorr.html

City and Town Population Totals: 2020-2021.
Data set: Annual Estimates of the Resident Population for Incorporated
Places of 50,000 or More, Ranked by July 1, 2021 Population:
April 1, 2020 to July 1, 2021.
Web page: https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-cities-and-towns.html

City and Town Population Totals: 2010-2019
Data set: Annual Estimates of the Resident Population for Incorporated
Places of 50,000 or More, Ranked by July 1, 2019 Population:
April 1, 2010 to July 1, 2019.
Web page: https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-cities-and-towns.html
}
}
\examples{
city_county_proxies_fun(n_cities=50, year=2020, rule="city", min_p_city=0.05)
city_county_proxies_fun(n_cities=25, year=2010, rule="county", min_p_county=0.1)
city_county_proxies_fun(n_cities=60, year=2010, rule="and", min_p_county=0.025, min_p_city=0.15)
}
