% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tract_zcta_cw.R
\name{load_tract_zcta_cw}
\alias{load_tract_zcta_cw}
\title{Load a census tract-to-ZCTA crosswalk}
\usage{
load_tract_zcta_cw(year = 2010)
}
\arguments{
\item{year}{a numeric year, either 2000 or 2010}
}
\value{
A data.frame containing full crosswalk details between census tract
and ZCTA.

The crosswalk contains one row per combination of overlapping tract and ZCTA.

Columns in the returned data are:
\itemize{
\item \code{state_fips} - (character) 2-digit state FIPS code .
\item \code{county_fips} - (character) 5-digit county FIPS code (2-digit state +
3 digit county).
\item \code{tract_fips} - (character) 11-digit census tract FIPS code (2-digit state +
3 digit county + 6 digit tract).
\item \code{zcta} - (character) 5-digit ZCTA (ZIP Code Tabulation Area) ID
\item \code{pop} - (numeric) population count estimate within the overlapping portion
of ZCTA and tract.
\item \code{p_tract} - (numeric) the proportion (valued 0-1) of the census tract
population residing in the portion which overlaps with the ZCTA.
\item \code{p_zcta} - (numeric) the porportion (valued 0-1) of the ZCTA population
residing in the portion which overlaps with the census tract.
}
}
\description{
Function to load in a data.frame containing a crosswalk between US census
tract and ZCTA (ZIP Code Tabulation Area) for year 2000 or 2010.
}
\details{
\subsection{Data sources used:}{

Missouri Census Data Center: Geographic Correspondence Engine.
Web page: https://mcdc.missouri.edu/applications/geocorr.html
}
}
\examples{
tract_zcta_2010_cw <- load_tract_zcta_cw(2010)

}
